+++ crates/azure_devops_types/src/azure_devops_license_kind.rs
@@ -3,7 +3,7 @@ use serde::Serialize;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Hash)]
 #[serde(rename_all = "PascalCase")]
-pub enum AzureDevOpsLicenseKind {
+pub enum AzureDevOpsLicenseType {
     /// Express means "Basic" in the UI
     #[serde(rename = "Account-Express")]
     #[serde(alias = "Basic")]
@@ -33,118 +33,129 @@ pub enum AzureDevOpsLicenseKind {
     #[serde(rename = "Msdn-Professional")]
     MsdnProfessional,
 
+    None,
+
     #[serde(untagged)]
     Other(String),
 }
 
-impl AzureDevOpsLicenseKind {
-    pub const VARIANTS: &'static [AzureDevOpsLicenseKind] = &[
-        AzureDevOpsLicenseKind::AccountExpress,
-        AzureDevOpsLicenseKind::AccountStakeholder,
-        AzureDevOpsLicenseKind::AccountAdvanced,
-        AzureDevOpsLicenseKind::MsdnEligible,
-        AzureDevOpsLicenseKind::MsdnEnterprise,
-        AzureDevOpsLicenseKind::MsdnProfessional,
+impl AzureDevOpsLicenseType {
+    pub const VARIANTS: &'static [AzureDevOpsLicenseType] = &[
+        AzureDevOpsLicenseType::AccountExpress,
+        AzureDevOpsLicenseType::AccountStakeholder,
+        AzureDevOpsLicenseType::AccountAdvanced,
+        AzureDevOpsLicenseType::MsdnEligible,
+        AzureDevOpsLicenseType::MsdnEnterprise,
+        AzureDevOpsLicenseType::MsdnProfessional,
     ];
 
     /// https://azure.microsoft.com/en-us/pricing/details/devops/azure-devops-services/
     pub fn cost_per_month_cad(&self) -> f64 {
         match self {
-            AzureDevOpsLicenseKind::AccountExpress => 8.30,
-            AzureDevOpsLicenseKind::AccountStakeholder => 0.0,
-            AzureDevOpsLicenseKind::AccountAdvanced => 71.93,
-            AzureDevOpsLicenseKind::MsdnEnterprise => 0.00,
-            AzureDevOpsLicenseKind::MsdnProfessional => 0.00,
-            AzureDevOpsLicenseKind::MsdnEligible => 0.00,
-            AzureDevOpsLicenseKind::Other(_) => 0.0,
+            AzureDevOpsLicenseType::AccountExpress => 8.30,
+            AzureDevOpsLicenseType::AccountStakeholder => 0.0,
+            AzureDevOpsLicenseType::AccountAdvanced => 71.93,
+            AzureDevOpsLicenseType::MsdnEnterprise => 0.00,
+            AzureDevOpsLicenseType::MsdnProfessional => 0.00,
+            AzureDevOpsLicenseType::MsdnEligible => 0.00,
+            AzureDevOpsLicenseType::None => 0.00,
+            AzureDevOpsLicenseType::Other(_) => 0.0,
         }
     }
 }
 
-impl std::str::FromStr for AzureDevOpsLicenseKind {
+impl std::str::FromStr for AzureDevOpsLicenseType {
     type Err = eyre::Error;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         // Use serde to leverage existing deserialization logic for the enum
         let q = format!("\"{}\"", s);
-        let license = serde_json::from_str::<AzureDevOpsLicenseKind>(&q)?;
+        let license = serde_json::from_str::<AzureDevOpsLicenseType>(&q)?;
         Ok(license)
     }
 }
 
-impl std::fmt::Display for AzureDevOpsLicenseKind {
+impl std::fmt::Display for AzureDevOpsLicenseType {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
-            AzureDevOpsLicenseKind::AccountExpress => write!(f, "Account-Express (Basic)"),
-            AzureDevOpsLicenseKind::AccountStakeholder => write!(f, "Account-Stakeholder"),
-            AzureDevOpsLicenseKind::AccountAdvanced => write!(f, "Account-Advanced (Basic+Test)"),
-            AzureDevOpsLicenseKind::MsdnEligible => write!(f, "Msdn-Eligible"),
-            AzureDevOpsLicenseKind::MsdnEnterprise => write!(f, "Msdn-Enterprise"),
-            AzureDevOpsLicenseKind::MsdnProfessional => write!(f, "Msdn-Professional"),
-            AzureDevOpsLicenseKind::Other(s) => write!(f, "{}", s),
+            AzureDevOpsLicenseType::AccountExpress => write!(f, "Account-Express (Basic)"),
+            AzureDevOpsLicenseType::AccountStakeholder => write!(f, "Account-Stakeholder"),
+            AzureDevOpsLicenseType::AccountAdvanced => write!(f, "Account-Advanced (Basic+Test)"),
+            AzureDevOpsLicenseType::MsdnEligible => write!(f, "Msdn-Eligible"),
+            AzureDevOpsLicenseType::MsdnEnterprise => write!(f, "Msdn-Enterprise"),
+            AzureDevOpsLicenseType::MsdnProfessional => write!(f, "Msdn-Professional"),
+            AzureDevOpsLicenseType::None => write!(f, "None"),
+            AzureDevOpsLicenseType::Other(s) => write!(f, "{}", s),
         }
     }
 }
 
 #[cfg(test)]
 mod license_tests {
-    use super::AzureDevOpsLicenseKind;
+    use super::AzureDevOpsLicenseType;
 
     #[test]
     pub fn deserializes_account_express() -> eyre::Result<()> {
-        let license = serde_json::from_str::<AzureDevOpsLicenseKind>(r#""Account-Express""#)?;
-        assert_eq!(license, AzureDevOpsLicenseKind::AccountExpress);
+        let license = serde_json::from_str::<AzureDevOpsLicenseType>(r#""Account-Express""#)?;
+        assert_eq!(license, AzureDevOpsLicenseType::AccountExpress);
         Ok(())
     }
 
     #[test]
     pub fn deserializes_account_stakeholder() -> eyre::Result<()> {
-        let license = serde_json::from_str::<AzureDevOpsLicenseKind>(r#""Account-Stakeholder""#)?;
-        assert_eq!(license, AzureDevOpsLicenseKind::AccountStakeholder);
+        let license = serde_json::from_str::<AzureDevOpsLicenseType>(r#""Account-Stakeholder""#)?;
+        assert_eq!(license, AzureDevOpsLicenseType::AccountStakeholder);
         Ok(())
     }
 
     #[test]
     pub fn deserializes_account_advanced() -> eyre::Result<()> {
-        let license = serde_json::from_str::<AzureDevOpsLicenseKind>(r#""Account-Advanced""#)?;
-        assert_eq!(license, AzureDevOpsLicenseKind::AccountAdvanced);
+        let license = serde_json::from_str::<AzureDevOpsLicenseType>(r#""Account-Advanced""#)?;
+        assert_eq!(license, AzureDevOpsLicenseType::AccountAdvanced);
         Ok(())
     }
 
     #[test]
     pub fn deserializes_msdn_eligible() -> eyre::Result<()> {
-        let license = serde_json::from_str::<AzureDevOpsLicenseKind>(r#""Msdn-Eligible""#)?;
-        assert_eq!(license, AzureDevOpsLicenseKind::MsdnEligible);
+        let license = serde_json::from_str::<AzureDevOpsLicenseType>(r#""Msdn-Eligible""#)?;
+        assert_eq!(license, AzureDevOpsLicenseType::MsdnEligible);
         Ok(())
     }
 
     #[test]
     pub fn deserializes_msdn_enterprise() -> eyre::Result<()> {
-        let license = serde_json::from_str::<AzureDevOpsLicenseKind>(r#""Msdn-Enterprise""#)?;
-        assert_eq!(license, AzureDevOpsLicenseKind::MsdnEnterprise);
+        let license = serde_json::from_str::<AzureDevOpsLicenseType>(r#""Msdn-Enterprise""#)?;
+        assert_eq!(license, AzureDevOpsLicenseType::MsdnEnterprise);
         Ok(())
     }
 
     #[test]
     pub fn deserializes_msdn_professional() -> eyre::Result<()> {
-        let license = serde_json::from_str::<AzureDevOpsLicenseKind>(r#""Msdn-Professional""#)?;
-        assert_eq!(license, AzureDevOpsLicenseKind::MsdnProfessional);
+        let license = serde_json::from_str::<AzureDevOpsLicenseType>(r#""Msdn-Professional""#)?;
+        assert_eq!(license, AzureDevOpsLicenseType::MsdnProfessional);
+        Ok(())
+    }
+
+    #[test]
+    pub fn deserializes_none() -> eyre::Result<()> {
+        let license = serde_json::from_str::<AzureDevOpsLicenseType>(r#""None""#)?;
+        assert_eq!(license, AzureDevOpsLicenseType::None);
         Ok(())
     }
 
     #[test]
     pub fn deserializes_other() -> eyre::Result<()> {
-        let license = serde_json::from_str::<AzureDevOpsLicenseKind>(r#""Custom-License""#)?;
+        let license = serde_json::from_str::<AzureDevOpsLicenseType>(r#""Custom-License""#)?;
         assert_eq!(
             license,
-            AzureDevOpsLicenseKind::Other("Custom-License".to_string())
+            AzureDevOpsLicenseType::Other("Custom-License".to_string())
         );
         Ok(())
     }
     #[test]
     pub fn display_account_express() -> eyre::Result<()> {
         assert_eq!(
-            format!("{}", AzureDevOpsLicenseKind::AccountExpress),
+            format!("{}", AzureDevOpsLicenseType::AccountExpress),
             "Account-Express (Basic)"
         );
         Ok(())
@@ -153,7 +164,7 @@ mod license_tests {
     #[test]
     pub fn display_account_advanced() -> eyre::Result<()> {
         assert_eq!(
-            format!("{}", AzureDevOpsLicenseKind::AccountAdvanced),
+            format!("{}", AzureDevOpsLicenseType::AccountAdvanced),
             "Account-Advanced (Basic+Test)"
         );
         Ok(())